alertWireframeを追加, 部屋番号送信後の処理を追加
