+++ QuizButton/Entrance/EntranceViewController.swift
@@ -22,22 +22,17 @@ class EntranceViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        viewModel = EntranceViewModel(sendButtonTap: sendButton.rx.tap.asSignal())
+        viewModel = EntranceViewModel(
+            dependency: (
+                EntranceWireframe(self),
+                AlertWireframe(self)
+            ),
+            sendButtonTap: sendButton.rx.tap.asSignal()
+        )
         
         roomNumberTextField.rx.text.orEmpty
             .bind(to: viewModel.roomNumberText)
             .disposed(by: disposeBag)
-        
-        viewModel.isAppropriateRoomNumber.emit(onNext: { isAppropriateRoomNumber in
-            if isAppropriateRoomNumber {
-                print("適切な部屋番号です")
-            } else {
-                // TODO: バリデーションのアラート表示
-                AlertUtility.showSingleAlert(title: "不適切な部屋番号です", message: "", viewController: self) { _ in
-                    self.roomNumberTextField.text = ""
-                }
-            }
-        }).disposed(by: disposeBag)
         
     }
